requirements.txt
pytube
ffmpeg-python
SpeechRecognition
pydub
soundfile
python-dotenv
datasets
nltk

